refactor: 랜덤 숫자 생성 부분 테스트 위치 알맞게 이동
