+++ src/test/java/racingcar/CarTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.number_generator.RandomNumberGenerator;
 import racingcar.number_generator.SpecificNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -42,14 +41,6 @@ public class CarTest {
         car.forward();
 
         assertThat(car.getDistance()).isEqualTo(1);
-    }
-
-    @Test
-    @DisplayName("0에서 9 사이의 랜덤한 숫자를 반환한다")
-    public void 영에서_9_사이의_랜덤한_숫자를_반환한다() throws Exception {
-        RandomNumberGenerator randomGenerator = new RandomNumberGenerator();
-        int number = randomGenerator.generate();
-        assertThat(number).isBetween(0, 9);
     }
 
     @ParameterizedTest
